Adds meaningful class names and max-width to .child
This commit aims to change the `.parent` and `.child` classes in order to make them more specific to the plugin. These terms are too generic and my cause conflicts with other stylesheets.

Also, `max-width: 100%;` is added to the `.child` class. This would avoid situations where the whole container is larger than the image size and the aspect ration must be preserved (a Masonry container, for example).